Move ingress test to its own file
